fix: Ball animation collision detection [PT-187271307]
This centralizes the collision detection and fixes a boundary error in the previous calculation.

NOTE: this also disables the final css translation of the balls to animate to their starting positions. This will be reenabled when the animation loop is updated at which time we will know when the animation completely stops.  The current code calls moveBackToInitialPositions at the end of each sample instead of at the end of all the samples.