fix: Mixer ball and spinner labels cannot be edited in Safari [PT-187127774]
It turns out Safari doesn't support either pointer styling or click events on svg text elements.  This commit hosts the click handler and styling to the circle element of the ball and adds an invisible rect behind the text for the spinner that acts as a click and styling source.