fix: Restore formula to just saving the value [PT-187221915]
The change in the work that validated the formula caused an object to be output for the formula with the value and the valid flag. This caused old files not to be loaded without a migration.  Since the valid flag can be computed there is no need to store it so the formula save state was returned to just be the value.