fix: Update spinner wedge dragging [PT-187636438]
This moves the drag handler to the edges between wedges and creates a wide click/hover path under the visible path to make it easier to select.  This also updates the visible path to be wider on hover/select and adds a center round element to let the user know the edge is draggable.