feat: Refactor animation [PT-187271348]
This refactors how the animation is handled in the plugin and moves the CODAP functions out of a hook and into a helpers file.

The main refactor is having the components per device register themselves when they mount with a callback function that is invoked when that device needs to animate.  This reduces the complexity of the components and allows the animation system to be able to pause and stop the animations.

The callback is called with the current animation settings that the component can use for animation.