fix: Allow device name input to resize [SAMPLER-6]
This changes the device name input to a textarea that defaults to 1 row but auto resizes to match the input.

This also fixes a bug that I noticed where users could clear the input field and set the variable name to the empty string.  Now if the user tries to do that and "blurs" the input to save it the column name is reset.