fix: Ensure minimum plugin window size [SAMPLER-19]
This ensures the plugin window is at least the minimum dimensions at startup.  This is a bit of a hack to get around a complex set of issues between CODAP v2 and v3 not currently having a way to tell the plugin the initial window size.

This also fixes the header styling so the horizontal scrollbar is not visible due to a padding overflow.