fix: Use minimum dimensions on new plugin instances [SAMPLER-86]
CODAP v3 sets a wider starting window for plugins.  This fix reduces the plugin to the minimum dimensions on a new plugin load, otherwise it ensures the plugin has the minimum dimensions when loaded from a saved document.

This is a noop in CODAP v2 as its plugin window is automatically sized smaller on initial load.