feat: Multiple sampler plugins in a document should be sequentially numbered [SAMPLER-78]
This uses a CODAP global variable, protected by a web lock, to set the instance count of a loaded sampler plugin.  The instance is then saved to state.

The instance is used in the plugin title and plug data context title.  When the instance is 1 nothing extra is output.  When the instance is > 1 then the instance value is shown in the plugin and data context title.