fix: Remove img tag from custom select
This removes the unneeded dropdown-arrow-icon.svg image tag from the custom select as it is already added in the css using an &::after rule.  Before this change there was a 404 message showing as the img src could not be resolved.  Padding has been added to take the place of the (broken) image.