feat: Added start of translation support [SAMPLER-93]
This imports the pull-strings.js script from the CODAP data interactives repo and simplifies it a bit to save the translations in a json file that is then loaded at startup in the locale manager.  The existing strings.json file that was added when the v2 plugin was created was removed and the new filename ("pulled-strings.json") is an attempt to show that the file should not be manually updated but rather updated via the script.

The plugin title and top level header component have been updated to use the tr() function in the locale manager as a proof of concept that it works.  As part of that the styling of the top level header was changed to remove the exact widths of elements as they need to grow based on the translated string length.