[Sampler-101] Feature: About pane in markdown
In prior versions the About pane was formatted using JSX. In version 0.44 the text came from a collection of awkward POEditor strings. In version 1.0.0 the POEditor strings were no longer used. With this PR the text in the About pane comes from a Markdown file. The English version replicates the text that occurred previously including the formatting.

The `react-markdown` npm module is installed to handle rendering of the About pane.

ChatGPT was used to translate the English text into the 9 additional languages for which there are POEditor strings. All the markdown files are contained in the `public` folder at the root of the repository. These translations are to be viewed as drafts that can be given to localizers for modification.

The new function `getAboutMarkdown` in localeManager.tsx respects the currently defined locale to determine which markdown file to use as the source of the text.

Locally, webpack.config.js was modified to include this `public` folder in its definition of `static`
      static: [
        { directory: path.resolve(__dirname, 'dist') },
        { directory: path.resolve(__dirname, 'public') },
      ],

It's not clear that this is the correct placement for deployment.